+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 
 ## [1.0.2] - 2025-06-19
 **Internal changes:**
+- Added a separate, a little bit advanced caching for this mod
 - Internal GUI code changes, the mod now uses [HelloGUI](https://github.com/tracystacktrace/HelloGUI) for dynamic config GUIs
 - Added a nicer and more advanced API for this mod. I can easily add my own additional description for specific items, so you do
 - Deleted an old event "based" description collector. Use other method provided in readme.md
 
 **Visible changes:**
+- Now you can see ID/Metadata of the hovered block. Open config screen to turn on it
 - Added settings to change X and Y offsets of a hint ui
-- Added a hint about magma blocks
+- Added additional description for following blocks: pistons, magma blocks, gears, carved pumpkings and watermelons, dungeon chest, cake and pie.
 
 ## [1.0.1] - 2025-06-13
 - Small bugfix: includes all item blocks that weren't added in version 1.0